backend/requirements.txt
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
qdrant-client>=1.16.0
cohere>=5.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
